% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget.R
\name{prec_budget}
\alias{prec_budget}
\title{Sample size or precision for a given budget}
\usage{
prec_budget(
  budget,
  price,
  sd = NULL,
  n = NULL,
  fail = 0,
  conf.width = NULL,
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{budget}{total research budget available.}

\item{price}{cost per sample.}

\item{sd}{standard deviation.}

\item{n}{number of observations.}

\item{fail}{proportion of samples expected to fail.}
}
\value{
Object of class "presize", a list with total \code{budget}, \code{price} per sample, \code{n} sample size,
 augmented with method and note elements.
}
\description{
\code{prec_budget} returns the sample size or the precision for the provided
budget and cost per sample.
}
\details{
Exactly one of the parameters \code{n} or \code{conf.width} must be passed as NULL,
and that parameter is determined from the other.

The precision is defined as the full width of the confidence interval. The
confidence interval calculated as \eqn{t(n - 1) * sd / sqrt(n)}, with t(n-1)
from the t-distribution with n-1 degrees of freedom.

\code{\link[stats]{uniroot}} is used to solve \code{n}.
}
\examples{
prec_budget(budget = 5000, price = 250, n = 20)
prec_budget(budget = 1000, price = 42, conf.width = 2.34)  # approximately the inverse of above
}
