% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_measures.R
\name{prec_icc}
\alias{prec_icc}
\title{Sample size or precision for an intraclass correlation}
\usage{
prec_icc(rho, k, n = NULL, conf.width = NULL, conf.level = 0.95)
}
\arguments{
\item{rho}{Desired intraclass correlation.}

\item{k}{number of observations per n (subject).}

\item{conf.width}{precision (the full width of the conficende interval)}

\item{conf.level}{confidence level}
}
\description{
\code{prec_icc} returns the sample size or the precision for the given
intraclass correlation
}
\details{
Exactly one of the parameters \code{n, conf.width} must be passed as NULL,
and that parameter is determined from the other.

Sample size or precision is calculated according to formula 3 in Bonett
(2002), which is an approximation. Whether ICC is calculated for a one-way or
a two-way ANOVA does not matter in the approximation. As suggested by the
author, \eqn{5*rho} is added to n, if \eqn{k = 2} and \eqn{rho \ge 7}.

n is rounded up to the next whole number using \code{ceiling}.
}
\references{
Bonett DG (2002). \emph{Sample size requirements for estimating
  intraclass correlations with desired precision}. Statistics in Medicine,
  21:1331-1335. \href{https://doi.org/10.1002/sim.1108}{doi:10.1002/sim.1108}
}
