% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{prec_rate}
\alias{prec_rate}
\title{Sample size or precision for a rate}
\usage{
prec_rate(r = NULL, t = 1, prec = NULL, conf.level = 0.95,
  method = c("score", "vs", "wald"))
}
\arguments{
\item{r}{rate or rate ratio. A vector of length one.}

\item{t}{time base for event count. A vector of length one.}

\item{prec}{precision (half the width of the conficende interval)}

\item{conf.level}{confidence level}

\item{method}{The method to use to calculate precision. Exactly one method
may be provided. Methods can be abbreviated.}
}
\value{
Object of class "power.htest", a list of arguments (including the
  computed one) augmented with method and note elements.
}
\description{
\code{prec_rate} returns the sample size or the precision for the provided
rate
}
\details{
Exactly one of the parameters \code{r, prec} must be passed as NULL, and that
parameter is determined from the other.

Currently, the score, variance stabilizing (vs), and wald method are
implemented to calculate the rate and the precision. If you would like an
exact conficence interval, use \code{\link[stats]{poisson.test}}. For small
rs (<5), the exact method is recommended.

If more than one method is specified or the method is miss-specified, the
'score' method will be used.
}
\references{
Barker, L. (2002) \emph{A Comparison of Nine Confidence Intervals
for a Poisson Parameter When the Expected Number of Events is \eqn{\le} 5},
The American Statistician, 56:2, 85-89,
\href{https://doi.org/10.1198/000313002317572736}{DOI:
10.1198/000313002317572736}
}
\seealso{
\code{\link[stats]{poisson.test}}
}
