% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_spec.R
\name{prec_sens_spec}
\alias{prec_sens_spec}
\title{Sample size or precision for sensitivity and specificity with known
prevalence}
\usage{
prec_sens_spec(sens, spec, prev, n = NULL, conf.width = NULL,
  conf.level = 0.95, tol = .Machine$double.eps^0.25)
}
\arguments{
\item{sens}{sensitivity}

\item{spec}{specificity}

\item{prev}{prevalence in population}

\item{n}{number of observations}

\item{conf.width}{precision (the full width of the conficende interval)}

\item{conf.level}{confidence level}

\item{tol}{numerical tolerance used in root finding, the default providing
(at least) four significant digits}
}
\description{
Exactly one of the parameters \code{n, conf.width} must be passed as NULL,
and that parameter is determined from the other.
}
\details{
When calculating sample sizes, the returned sample size is for both
\code{sens} and \code{spec} to have confidence intervals at most
\code{conf.width} wide, given the other parameters.

When calulating sample precisions, the returned value is the widest
confidence interval that both \code{sens} and \code{spec} could take, given
the other parameters.

This function is suitable for cases where the disease state is unknown at the
time of sampling (e.g. a prospective study with consecutive patients).
}
\examples{
# calculate number of observations with sensitivity of 90\%, specificity of
# 85\%, prevalence of 20\%, confidence interval around \\code{sens} or
# \\code{spec} of 10\%
prec_sens_spec(.9, .85, prev = .2, conf.width = .1)
# calculate confidence interval width at sensitivity of 90\%, specificity of
# 85\%, prevalence of 20\%, 173 observations
prec_sens_spec(.9, .85, prev = .2, n = 173)
# multiple values
sens <- c(.9, .85)
spec <- c(.85, .75)
prec_sens_spec(sens, spec, prev = .2, conf.width = .1)
prec_sens_spec(sens, spec, prev = .2, n = 245)
}
\references{
Buderer, N.M.F. (1996) \emph{Statistical Methodology: I.
Incorporating the Prevalence of Disease into the Sample Size Calculation for
Sensitivity and Specificity}, Academic Emergency Medicine, 3:9, 895-900, \href{https://doi.org/10.1111/j.1553-2712.1996.tb03538.x}{DOI: 10.1111/j.1553-2712.1996.tb03538.x}
}
