% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{prec_auc}
\alias{prec_auc}
\title{Sample size or precision for AUC}
\usage{
prec_auc(auc, prev, n = NULL, conf.width = NULL, conf.level = 0.95,
  ...)
}
\arguments{
\item{auc}{AUC value}

\item{prev}{prevalence}

\item{n}{number of observations}

\item{conf.width}{precision (the full width of the confidence interval)}

\item{conf.level}{confidence level}

\item{...}{other arguments to \code{optimize}}
}
\description{
Calculate the sample size from AUC, prevalence and confidence interval width
or the expected confidence interval width from AUC, prevalence and sample
size
}
\details{
Sample size is derived by optimizing the difference between the difference
between the lower and upper limits of the confidence interval and
\code{conf.width}.
}
