% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{prec_mean}
\alias{prec_mean}
\title{Sample size or precision for a mean}
\usage{
prec_mean(mu, sd, n = NULL, prec = NULL, conf.level = 0.95,
  tol = .Machine$double.eps^0.25)
}
\arguments{
\item{mu}{mean}

\item{sd}{standard deviation}

\item{n}{number of observations}

\item{prec}{precision (half the width of the conficende interval)}

\item{conf.level}{confidence level}

\item{tol}{numerical tolerance used in root finding, the default providing
(at least) four significant digits}
}
\value{
Object of class "presize", a list with
  \describe{
    \item{mu}{mean}
    \item{sd}{standard deviation}
    \item{n}{sample size}
    \item{prec}{precision (half the width of the conficence interval)}
    \item{lwr}{lower end of confidence interval}
    \item{upr}{higher end of confidence interval}
  } augmented with method and note elements.
}
\description{
\code{prec_mean} returns the sample size or the precision for the provided
mean and standard deviation
}
\details{
Exactly one of the parameters \code{n, prec} must be passed as NULL, and that
parameter is determined from the other. The confidence interval calculated as
\eqn{t(n - 1) * sd / sqrt(n)}, with t(n-1) from the t-distribution with n-1
degrees of freedom.

The function uses \code{\link[base]{expand.grid}} to provide an estimate of n
or prec for every possible combination of supplied arguments.

\code{\link[stats]{uniroot}} is used to solve n for the score and
exact method. Agresti-coull can be abbreviated by ac.
}
\examples{
prec_mean(mu = 5, sd = 2.5, n = 20)
prec_mean(mu = 5, sd = 2.5, prec = 1.1)  # approximately the inverse of above

# Expands arguments
prec_mean(mu = c(5, 10, 15), sd = c(2.5, 5), n = 1:2 * 10)

}
