% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{prec_mean}
\alias{prec_mean}
\title{Sample size or precision for a mean}
\usage{
prec_mean(mu, sd, n = NULL, prec = NULL, conf.level = 0.95)
}
\arguments{
\item{mu}{mean. A vector of length one.}

\item{sd}{standard deviation. A vector of length one.}

\item{n}{number of observations}

\item{prec}{precision (half the width of the conficende interval)}

\item{conf.level}{confidence level}
}
\value{
Object of class "power.htest", a list with
  \describe{
  \item{mu}{mean}
  \item{sd}{standard deviation}
  \item{n}{sample size}
  \item{prec}{precision (half the width of the conficence interval)}
  \item{lwr}{lower end of confidence interval}
  \item{upr}{higher end of confidence interval}
  } augmented with method and note elements.
}
\description{
\code{prec_mean} returns the sample size or the precision for the provided
mean and standard deviation
}
\details{
Exactly one of the parameters \code{n, prec} must be passed as NULL, and that
parameter is determined from the other.

The confidence interval calculated as \eqn{z * sd / sqrt(n)}, with z from
the standard normal distribution.
}
\examples{
prec_mean(mu = 5, sd = 2.5, n = 20)
prec_mean(mu = 5, sd = 2.5, prec = 1.1)
}
