% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{prec_prop}
\alias{prec_prop}
\title{Sample size or precision for a proportion}
\usage{
prec_prop(p, n = NULL, prec = NULL, conf.level = 0.95,
  method = c("wald", "ac", "agresti-coull", "wilson"),
  tol = .Machine$double.eps^0.25)
}
\arguments{
\item{p}{proportion}

\item{n}{number of observations}

\item{prec}{precision (half the width of the conficende interval)}

\item{conf.level}{confidence level}

\item{method}{The method to use to calculate sample size or precision.
Exactly one method may be provided. Methods can be abbreviated.}

\item{tol}{numerical tolerance used in root finding, the default providing
(at least) four significant digits}
}
\value{
Object of class "power.htest", a list of arguments (including the
  computed one) augmented with method and note elements.
}
\description{
\code{prec_prop} returns the sample size or the precision for the provided
proportion
}
\details{
Exactly one of the parameters \code{n, prec} must be passed as NULL, and that
parameter is determined from the other.

Currently, the wald, agresti-coull, and wilson method are implemented.
\code{\link[stats]{uniroot}} is used to solve n for the agresti-coull and
wilson method.
}
\examples{
add(1, 1)
add(10, 1)
}
\seealso{
\code{\link[stats]{binom.test}} \code{\link[binom]{binom.confint}} \code{\link[Hmisc]{binconf}}
}
