% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differences.R
\name{prec_riskdiff}
\alias{prec_riskdiff}
\title{Sample size or precision for risk difference}
\usage{
prec_riskdiff(p1, p2, n1 = NULL, conf.width = NULL, r = 1,
  conf.level = 0.95, method = c("newcombe"),
  tol = .Machine$double.eps^0.25)
}
\arguments{
\item{p1}{Risk among unexposed}

\item{p2}{Risk among exposed}

\item{n1}{Number of patients in unexposed group}

\item{conf.width}{precision (the full width of the conficende interval)}

\item{r}{allocation ratio (relative size of unexposed and exposed cohort
(\code{n1} / \code{n2}))}

\item{conf.level}{confidence level}

\item{method}{The method to use to calculate precision. Exactly one method
may be provided. Methods can be abbreviated.}

\item{tol}{numerical tolerance used in root finding, the default providing
(at least) four significant digits}
}
\description{
\code{prec_riskdiff} returns the risk difference and the sample size or the
precision for the provided proportions
}
\details{
Exactly one of the parameters \code{n1, conf.width} must be passed as NULL,
and that parameter is determined from the other.

Newcombe is implemented as equation 10 in Newcombe (1998)

\code{\link[stats]{uniroot}} is used to solve n for the newcombe method.

The function uses \code{\link[base]{expand.grid}} to provide an estimate of n
or conf.width for every possible combination of supplied arguments.
}
\examples{
# Validate Newcombe (1998)
prec_riskdiff(p1 = 56/70, p2 = 48/80, n1 = 70, r = 70/80, met = "newcombe")  # Table IIa
prec_riskdiff(p1 = 10/10, p2 = 0/10, n1 = 10, met = "newcombe")  # Table IIh
}
\references{
Newcombe RG (1998) \emph{Interval estimation for the difference
  between independent proportions: comparison of eleven methods}, Statistics
  in Medicine, 17:873-890
}
