% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presize-package.R
\docType{package}
\name{presize-package}
\alias{presize}
\alias{presize-package}
\title{presize: Precision Based Sample Size Calculation}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Bland (2009) <doi:10.1136/bmj.b3985> recommended to
    base study sizes on the width of the confidence interval rather the power of 
    a statistical test. The goal of 'presize' is to provide functions for such 
    precision based sample size calculations. For a given sample size, the 
    functions will return the precision (width of the confidence interval), and 
    vice versa.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CTU-Bern/presize}
  \item \url{https://ctu-bern.github.io/presize/}
  \item Report bugs at \url{https://github.com/CTU-Bern/presize/issues}
}

}
\author{
\strong{Maintainer}: Alan G. Haynes \email{alan.haynes@ctu.unibe.ch}

Authors:
\itemize{
  \item Armando Lenz \email{armando.lenz@ctu.unibe.ch}
  \item Andreas Limacher \email{andreas.limacher@ctu.unibe.ch}
}

Other contributors:
\itemize{
  \item Odile Stalder \email{odile.stalder@ctu.unibe.ch} [contributor]
}

}
\keyword{internal}
