% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{prec_sens}
\alias{prec_sens}
\alias{prec_spec}
\title{Sample size and precision of sensitivity and specificity}
\usage{
prec_sens(sens, n = NULL, conf.width = NULL, conf.level = 0.95, ...)

prec_spec(spec, n = NULL, conf.width = NULL, conf.level = 0.95, ...)
}
\arguments{
\item{sens, spec}{proportions}

\item{n}{number of observations}

\item{conf.width}{precision (the full width of the conficende interval)}

\item{conf.level}{confidence level}

\item{...}{other arguments to uniroot (e.g. \code{tol})}
}
\value{
Object of class "presize", a list of arguments (including the
  computed one) augmented with method and note elements.
}
\description{
These functions act as wrappers for \code{prec_prop}.
}
\seealso{
\code{prec_prop}
}
